Debugging blarrg timing tests, comparing to mooneye.
TIMA doesn't seem to be incrementing correctly (not soon /often enough?) which is failing blarrg timer tests.

+++ src/console/console.rs
@@ -203,7 +203,7 @@ impl Console {
         self.cycles += self.cpu.handle_interrupts(&mut self.mmu);
 
         // TIMER
-        if self.timer.step(&mut self.mmu, self.cycles as u8) {
+        if self.timer.step(&mut self.mmu, self.cycles as u16) {
             self.cpu.interrupts.request(InterruptRegBit::Timer, &mut self.mmu);
         }
